flask backend/requirements.txt
flask 
requests
huggingface-hub
pillow